create searchparam config so people can use that module without using subscriptions
