+++ hapi-fhir-jpaserver-subscription/src/main/java/ca/uhn/fhir/jpa/subscription/module/config/BaseSubscriptionConfig.java
@@ -20,7 +20,6 @@ package ca.uhn.fhir.jpa.subscription.module.config;
  * #L%
  */
 
-import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.interceptor.executor.InterceptorService;
 import ca.uhn.fhir.jpa.subscription.module.cache.ISubscribableChannelFactory;
 import ca.uhn.fhir.jpa.subscription.module.cache.LinkedBlockingQueueSubscribableChannelFactory;
@@ -33,8 +32,6 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableScheduling
 @ComponentScan(basePackages = {"ca.uhn.fhir.jpa.searchparam", "ca.uhn.fhir.jpa.subscription.module"})
 public abstract class BaseSubscriptionConfig {
-	public abstract FhirContext fhirContext();
-
 	@Bean
 	public ISubscribableChannelFactory blockingQueueSubscriptionDeliveryChannelFactory() {
 		return new LinkedBlockingQueueSubscribableChannelFactory();
@@ -44,6 +41,4 @@ public abstract class BaseSubscriptionConfig {
 	public InterceptorService interceptorRegistry() {
 		return new InterceptorService("hapi-fhir-jpa-subscription");
 	}
-
-
 }
